Revert "The Companion trait did not work as expected"
This reverts commit 7167e54a9c2e0a089154d7cda469f33a0797fc4f.
